switched bc with awk- hence removing the bc dep
Instead of using "bc -l" in line 115 I switched it with awk. It was the only occurence of it. bc is a great tool and I have it installed however it is not common. One may save one extra dependency if using awk instead because it is already used a couple of times and therefore already must be installed.